Create Colton'sCommands2
I was testing to make sure the World coordinates were being parsed correctly and I made an error. In the Lon section of all of them for the lonMinutes and lonSeconds I had the substring reading the lat version of it instead of the Lon because I cried and pasted the logic and forgot to change that. the above commands for that should work